requirements.txt
requests==2.31.0
discord.py==1.7.3
mcrcon==0.7.0
python-dotenv==0.19.2
